@@ -1,6 +1,12 @@
-const getGenresHandler = (req, res) => {
+const { getAllGenres } = require("../controllers/genresControllers");
+
+const getGenresHandler = async (req, res) => {
   try {
-  } catch (error) {}
+    const genres = await getAllGenres();
+    res.status(200).json(genres);
+  } catch (error) {
+    res.status(400).json({ error: error.message });
+  }
 };
 
 module.exports = { getGenresHandler };
