@@ -1,8 +1,11 @@
-const Home = (props) => {
+import { Link } from "react-router-dom";
+const Home = () => {
   return (
     <div>
       <h1>Tu página de videojuegos</h1>
-      <button name="button">VAMOS</button>
+      <Link to="/page">
+        <button name="button">VAMOS</button>
+      </Link>
     </div>
   );
 };
